Address SonarCloud Quality Gate noise on the reliability PR
- Remove redundant (uint32_t) casts on already-uint32_t shifts (comms.cpp
  reverse_bytes, sectors; comms_legacy.cpp CRC32_val, toothHistory[x]).
  These cleared the original S3949 false positives but introduced S1905
  redundant-cast warnings, which net out for nothing.

- For the genuine uint8_t shift bug in the SD command path, declare the
  sector locals as uint32_t directly instead of casting at the use site.
  Same generated code, no S3949, no S1905.

- Initialise raw[] and value in readSerialIntegralTimeout to silence
  M23_321 (locals are written by readSerialTimeout / memcpy before use,
  but the explicit init makes the rule happy).

+++ speeduino/comms_legacy.cpp
@@ -561,9 +561,9 @@ void legacySerialHandler(byte cmd, Stream &targetPort, SerialStatus &targetStatu
         uint32_t CRC32_val = calculatePageCRC32( targetPort.read() );
         
         //Split the 4 bytes of the CRC32 value into individual bytes and send
-        targetPort.write( (uint8_t)(((uint32_t)CRC32_val >> 24) & 255U) );
-        targetPort.write( (uint8_t)(((uint32_t)CRC32_val >> 16) & 255U) );
-        targetPort.write( (uint8_t)(((uint32_t)CRC32_val >> 8) & 255U) );
+        targetPort.write( (uint8_t)((CRC32_val >> 24) & 255U) );
+        targetPort.write( (uint8_t)((CRC32_val >> 16) & 255U) );
+        targetPort.write( (uint8_t)((CRC32_val >> 8) & 255U) );
         targetPort.write( (uint8_t)(CRC32_val & 255U) );
         
         targetStatusFlag = SERIAL_INACTIVE;
@@ -1178,9 +1178,9 @@ void sendToothLog_legacy(byte startOffset) /* Blocking */
       serialStatusFlag = SERIAL_TRANSMIT_TOOTH_INPROGRESS_LEGACY; 
       for (uint8_t x = startOffset; x < _countof(toothHistory); ++x)
       {
-        primarySerial.write((uint8_t)((uint32_t)toothHistory[x] >> 24));
-        primarySerial.write((uint8_t)((uint32_t)toothHistory[x] >> 16));
-        primarySerial.write((uint8_t)((uint32_t)toothHistory[x] >> 8));
+        primarySerial.write((uint8_t)(toothHistory[x] >> 24));
+        primarySerial.write((uint8_t)(toothHistory[x] >> 16));
+        primarySerial.write((uint8_t)(toothHistory[x] >> 8));
         primarySerial.write((uint8_t)toothHistory[x]);
       }
       currentStatus.isToothLog1Full = false;
